chore: split config from app/server/app.js #242
